Use write temp and replace to avoid 2 processes working on the same xlf file
To use File.Replace need to update to netcoreapp2.0

Use FileShare.None in Document.Save(path) and FileShare.Read in Document.Load(path)

Replace xlfDoucment.Save with SaveXlfDocument helper that saves to temp path and does File.Replace in a retry loop to move the file.

Reply might be needed, however, when test on Roslyn, no concurrency issue happened.

This maximize parallelism and get rid of the need to create 1:1 mutex:file, eliminating the naming problem and other things we may be missing here such as handling AbandonedMutexException.
On Windows:
•	Readers will prevent the file from being overwritten due to FileShare.Read.
•	Readers can read in parallel, but when there's contention with a writer
On Unix:
•	FileShare.Read does nothing, but...
•	File.Replace is implemented with rename system call that will mean that even though writers can overwrite while
reading is happening, each reader will see file before or after overwrite, not in between, which is the same reader-writer semantic we want.